+++ Sources/Steps/Components/Item.swift
@@ -1,6 +1,6 @@
 //
-//  StepElement.swift
-//  curlyhair
+//  Item.swift
+//  Steps
 //
 //  Created by Saul Moreno Abril on 09/04/2020.
 //  Copyright © 2020 Saul Moreno Abril. All rights reserved.
@@ -9,16 +9,17 @@
 import SwiftUI
 import SwifterSwiftUI
 
-/// Element to represent each step
-struct StepElement: View {
-    /// Index of this step
-    private(set) var index: Int
+/// Item to represent each step
+@available(iOS 13.0, OSX 10.15, tvOS 13.0, watchOS 6.0, *)
+struct Item<Element>: View {
+    /// Get step object for the set index
+    var step: Step
 
     /// The main state
-    @ObservedObject private(set) var state: StepsState
+    @EnvironmentObject var state: StepsState<Element>
 
     /// The style of the component
-    @EnvironmentObject var config: StepsConfig
+    @EnvironmentObject var config: Config
 
     /// Helper to inspect
     let inspection = Inspection<Self>()
@@ -29,34 +30,24 @@ struct StepElement: View {
     /// Current offset to be animated
     @State private var offset: CGFloat = 0
 
-    /// Get current step state  for the set index
-    private var stepState: StepState {
-        return state.stepStateFor(index: index)
-    }
-
-    /// Get step object for the set index
-    private var step: Step {
-        return state.stepAtIndex(index: index)
+    /// Max diff offset to be animated
+    private var maxOffset: CGFloat {
+        return config.itemSpacing * 2
     }
 
     /// Get image for the current step
     private var image: Image? {
-        return stepState != .completed ? step.image : config.image
+        return step.state != .completed ? step.image : config.defaultImage
     }
 
     /// Get foreground color for the current step
     private var foregroundColor: Color {
-        switch stepState {
+        switch step.state {
         case .uncompleted:
             return config.disabledColor
         default:
             return config.primaryColor
         }
-    }
-
-    /// Max diff offset to be animated
-    private var maxOffset: CGFloat {
-        return config.spacing * 2
     }
 
     /// Update the offset to animate according the next index
@@ -67,28 +58,34 @@ struct StepElement: View {
             return
         }
 
-        if ((previousIndex == state.steps.endIndex && diff > 0) ||
-            (nextIndex == state.steps.endIndex && diff < 0)) {
+        if ((previousIndex == state.data.endIndex && diff > 0) ||
+            (nextIndex == state.data.endIndex && diff < 0)) {
             offset = 0
-        } else if (previousIndex == index) {
+        } else if (previousIndex == step.index) {
             offset = CGFloat(diff) * maxOffset
-        } else if(nextIndex == index) {
+        } else if(nextIndex == step.index) {
             offset = -CGFloat(diff) * maxOffset
         } else {
             offset = 0
         }
     }
 
+    private func onCompletionEffect() {
+        if self.offset != 0 {
+            offset = 0
+        }
+    }
+
     var body: some View {
-        StepContainer(title: step.title) {
+        Container(title: step.title) {
             ifLet(image, then: {
                 $0.resizable()
             }, else: {
-                Text("\(index + 1)").font(.system(size: config.size))
+                Text("\(step.index + 1)").font(.system(size: config.size))
             })
                 .frame(width: config.size, height: config.size)
                 .padding(config.figurePadding)
-                .if(index == state.currentIndex, then: {
+                .if(step.index == state.currentIndex, then: {
                     $0.background(config.primaryColor).foregroundColor(config.secondaryColor)
                 }, else: {
                     $0.overlay(
@@ -98,28 +95,32 @@ struct StepElement: View {
                 .cornerRadius(config.size)
         }
         .foregroundColor(foregroundColor)
-        .modifier(OffsetEffect(offset: offset, pct: abs(offset) > 0 ? 1 : 0))
-        .animation(config.animation)
-        .onReceive(state.$currentIndex, perform: { (nextIndex) in
-            self.updateOffset(nextIndex: nextIndex)
-            let previousOffset = self.offset
-            DispatchQueue.main.asyncAfter(deadline: .now() + self.config.animationDuration) {
-                if (self.offset != 0 && previousOffset == self.offset) {
-                    self.offset = 0
-                }
-            }
-            self.previousIndex = nextIndex
-        })
+        .modifier(
+            OffsetEffect(
+                offset: offset,
+                pct: abs(offset) > 0 ? 1 : 0,
+                onCompletion: onCompletionEffect
+            )
+        )
+            .animation(config.animation)
+            .onReceive(state.$currentIndex, perform: { (nextIndex) in
+                self.updateOffset(nextIndex: nextIndex)
+                self.previousIndex = nextIndex
+            })
             .onReceive(inspection.notice) { self.inspection.visit(self, $0) }
     }
 }
 
 #if DEBUG
-struct StepElement_Previews: PreviewProvider {
+struct Item_Previews: PreviewProvider {
     static var previews: some View {
-        let steps = [Step(title: "First"), Step(), Step()]
-        let state = StepsState(steps: steps)
-        return StepElement(index: 0, state: state)
+        let steps = ["First", "", ""]
+        let state = StepsState(data: steps)
+        return (
+            Steps(state: state, onCreateStep: { element in
+                Step(title: element)
+            }).padding()
+        )
     }
 }
 #endif