+++ Sources/Steps/Steps.swift
@@ -1,52 +1,56 @@
 import SwiftUI
 import Combine
+import SwifterSwiftUI
 
 /// 🏄‍♂️ A navigation bar that guides users through the steps of a task.
-public struct Steps: View {
+public struct Steps<Element>: View {
+
     /// The main state of the component
-    @ObservedObject public private(set) var state: StepsState
+    @ObservedObject private(set) var state: StepsState<Element>
+
+    /// Block to create each step
+    let onCreateStep: (Element) -> Step
 
     /// The style of the component
-    public private(set) var config: StepsConfig
+    let config = Config()
 
     /// Helper to inspect
     let inspection = Inspection<Self>()
 
-    /// Initializes a new steps component with the initial state and config.
+    /// Initializes a new `Steps`.
     ///
-    /// - Parameters:
-    ///   - state: Initial state.
-    ///   - config: Initial config.
-    public init(state: StepsState, config: StepsConfig = StepsConfig()) {
+    /// - Parameter state: State to manage component
+    /// - Parameter onCreateStep: Block to create each step
+    public init(state: StepsState<Element>, onCreateStep: @escaping (Element) -> Step) {
         self.state = state
-        self.config = config
+        self.onCreateStep = onCreateStep
     }
 
-    /// Initializes a new step element.
+    /// Initializes a new separator
     ///
     /// - Parameters:
     ///   - index: index of the step
-    private func makeStepAt(index: Int) -> some View {
-        return StepElement(index: index, state: state)
-    }
+    private func renderIndex(_ index: Int) -> some View {
+        let element = state.data[index]
+        var step = onCreateStep(element)
+        step.index = index
+        step.state = stateFor(step: step)
 
-    /// Initializes a new step element.
-    ///
-    /// - Parameters:
-    ///   - index: index of the step
-    private func makeSeparatorAt(index: Int) -> some View {
-        return StepSeparator(index: index, state: state)
+        let first = Item<Element>(step: step)
+        var second: Separator<Element>?
+        if (index < state.data.endIndex - 1) {
+            second = Separator(step: step)
+        }
+        return ViewBuilder.buildBlock(first,second)
     }
 
     public var body: some View {
-        HStack(alignment: .top, spacing: config.spacing) {
-            ForEach(state.steps.indices) { index in
-                self.makeStepAt(index: index)
-                if (index < self.state.steps.endIndex - 1) {
-                    self.makeSeparatorAt(index: index)
-                }
+        HStack(alignment: .top, spacing: config.itemSpacing) {
+            ForEach(state.data.indices) { index in
+                self.renderIndex(index)
             }
         }
+        .environmentObject(state)
         .environmentObject(config)
         .onReceive(inspection.notice) { self.inspection.visit(self, $0) }
     }
@@ -55,11 +59,85 @@ public struct Steps: View {
 #if DEBUG
 struct Steps_Previews: PreviewProvider {
     static var previews: some View {
-        let steps = [Step(title: "First"), Step(), Step()]
-        let state = StepsState(steps: steps)
+        let steps = ["First", "", ""]
+        let state = StepsState(data: steps)
         return (
-            Steps(state: state).padding()
+            Steps(state: state, onCreateStep: { element in
+                Step(title: element)
+            }).padding()
         )
     }
 }
 #endif
+
+// MARK: Builders
+extension Steps {
+    /// Adds space between each page
+    ///
+    /// - Parameter value: Spacing between elements
+    public func itemSpacing(_ value: CGFloat) -> Self {
+        config.itemSpacing = value
+        return self
+    }
+
+    /// Modify size of each step element
+    ///
+    /// - Parameter value: Size of each step element
+    public func size(_ value: CGFloat) -> Self {
+        config.size = value
+        return self
+    }
+
+    /// Modify the line thickness of all elements
+    ///
+    /// - Parameter value: Line thickness of all elements
+    public func lineThickness(_ value: CGFloat) -> Self {
+        config.lineThickness = value
+        return self
+    }
+
+    /// Modify color for current and completed steps
+    ///
+    /// - Parameter value: Color for current and completed steps
+    public func primaryColor(_ value: Color) -> Self {
+        config.primaryColor = value
+        return self
+    }
+
+    /// Modify color for text inside step element
+    ///
+    /// - Parameter value: Color for text inside step element
+    public func secondaryColor(_ value: Color) -> Self {
+        config.secondaryColor = value
+        return self
+    }
+
+    /// Modify color for text inside step element
+    ///
+    /// - Parameter value: Color for uncompleted steps
+    public func disabledColor(_ value: Color) -> Self {
+        config.disabledColor = value
+        return self
+    }
+
+    /// Modify color for text inside step element
+    ///
+    /// - Parameter value: Color for uncompleted steps
+    public func defaultImage(_ value: Image) -> Self {
+        config.defaultImage = value
+        return self
+    }
+}
+
+// MARK: Helpers
+extension Steps {
+    /// Get state for step at an index
+    func stateFor(step: Step) -> Step.State {
+        if (step.index < state.currentIndex) {
+            return .completed
+        } else if step.index == state.currentIndex {
+            return .current
+        }
+        return .uncompleted
+    }
+}