+++ Tests/StepsTests/StepsTests.swift
@@ -11,24 +11,26 @@ import ViewInspector
 @testable import Steps
 
 extension Steps: Inspectable { }
-extension TupleView : Inspectable where T == (StepElement, StepSeparator?) {}
+extension TupleView : Inspectable where T == (Item<String>, Separator<String>?) {}
 
 final class StepsTests: XCTestCase {
-    let config = StepsConfig()
-    let steps = [Step(), Step()]
+    let config = Config()
+    let data = ["First", "Second"]
     lazy var state: StepsState = {
-        return StepsState(steps: steps)
+        return StepsState(data: data)
     }()
 
     func testSteps() {
-        let container = Steps(state: state, config: config)
+        let container = Steps(state: state, onCreateStep: { string in
+            Step(title: string, image: Image(""))
+        })
         let exp = container.inspection.inspect { view in
-            let count = try view.actualView().state.steps.count
+            let count = try view.actualView().state.data.count
             for i in 0...count+1 {
                 if (i < count) {
-                    XCTAssertNoThrow(try view.hStack().forEach(0).view(TupleView<(StepElement, StepSeparator?)>.self, i))
+                    XCTAssertNoThrow(try view.hStack().forEach(0).view(TupleView<(Item<String>, Separator<String>?)>.self, i))
                 } else {
-                    XCTAssertThrowsError(try view.hStack().forEach(0).view(TupleView<(StepElement, StepSeparator?)>.self, i))
+                    XCTAssertThrowsError(try view.hStack().forEach(0).view(TupleView<(Item<String>, Separator<String>?)>.self, i))
                 }
             }
         }
@@ -36,7 +38,30 @@ final class StepsTests: XCTestCase {
         wait(for: [exp], timeout: 0.1)
     }
 
-    static var allTests = [
-        ("testSteps", testSteps)
-    ]
+    func testBuilders() {
+        let itemSpacing: CGFloat = 10.0
+        let size: CGFloat = 15.0
+        let lineThickness: CGFloat = 2.0
+        let primaryColor = Color.red
+        let secondaryColor = Color.yellow
+        let disabledColor = Color.gray
+        let defaultImage = Image("")
+
+        let container = Steps(state: state, onCreateStep: { string in Step(title: string)})
+            .itemSpacing(itemSpacing)
+            .size(size)
+            .lineThickness(lineThickness)
+            .primaryColor(primaryColor)
+            .secondaryColor(secondaryColor)
+            .disabledColor(disabledColor)
+            .defaultImage(defaultImage)
+
+        XCTAssertEqual(container.config.itemSpacing, itemSpacing)
+        XCTAssertEqual(container.config.size, size)
+        XCTAssertEqual(container.config.lineThickness, lineThickness)
+        XCTAssertEqual(container.config.primaryColor, primaryColor)
+        XCTAssertEqual(container.config.secondaryColor, secondaryColor)
+        XCTAssertEqual(container.config.disabledColor, disabledColor)
+        XCTAssertEqual(container.config.defaultImage, defaultImage)
+    }
 }