+++ CHANGELOG.md
@@ -17,6 +17,13 @@ The changelog for **Steps**. Also see the [releases](https://github.com/asam139/
 ### Security
 
 ---
+## [v0.3.4](https://github.com/asam139/Steps/releases/tag/0.3.4)
+### Added
+- New method to change the current step.
+### Change
+- Implement component following the builder pattern.
+- Improve animations.
+
 ## [v0.3.3](https://github.com/asam139/Steps/releases/tag/0.3.3)
 ### Fixed
 - Fix layout for title